Feat : 레이아웃 구성 변경
