+++ src/components/features/MyPage/MyInfo/MyInfo.styled.ts
@@ -2,9 +2,8 @@ import styled from '@emotion/styled';
 
 export const MyInfoContainer = styled.section`
 	width: 100%;
-	padding: 52px 32px 40px 32px;
+	padding: 72px 32px 40px 32px;
 	background-color: ${({ theme }) => theme.colors.bt_primary[20]};
-	margin-top: -20px;
 `;
 
 export const MyProfile = styled.div`
